refactor: separate database connection from DAO
